Update setting spatial attrs
